현재 위치와 가게 사이의 거리 구하기

+++ src/main/java/com/depromeet/team5/controller/StoreController.java
@@ -26,7 +26,7 @@ public class StoreController {
     }
 
     @GetMapping("/get")
-    public ResponseEntity<List<StoreCardDto>> getAll(@RequestParam Long latitude, @RequestParam Long longitude, @RequestParam Long radius) {
+    public ResponseEntity<List<StoreCardDto>> getAll(@RequestParam Float latitude, @RequestParam Float longitude, @RequestParam Float radius) {
         return new ResponseEntity<>(storeService.getAll(latitude, longitude, radius), HttpStatus.OK);
     }
